update programs for CS50 2022 fall

+++ harvard/cs50/01_c/truncation.c
@@ -1,12 +1,13 @@
+// Division with longs, demonstrating truncation
+
 #include ".vscode/cs50.h"
 #include <stdio.h>
 
 int main(void) 
 {
-
     // Get input from users
-    int x = get_int("x: ");
-    int y = get_int("y: ");
+    long x = get_int("x: ");
+    long y = get_int("y: ");
 
     // Divide x by y
     float z = x / y;
